# lib/api-zod/src/generated/types/uploadUrlRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
  /** Request an upload slot in the PUBLIC bucket (profile avatars, which must load in a plain <img> with no auth header). Defaults to false, which issues an auth-gated object path. Only set this for media that is genuinely safe to serve to anonymous callers. */
  isPublicAccess?: boolean;
}